+++ CHANGELOG.md
@@ -1,7 +1,19 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [0.3.0] -
+## [0.4.0] - 
+
+### Added
+- Grid with periodic boundary conditions
+- VTK Writer for distributed grid, now is the default writer
+
+### Fixed
+- GPU compilation
+
+### Changed
+
+
+## [0.3.0] - 16-04-2016
 
 ### Added
 - Molacular Dynamic example
@@ -17,7 +29,7 @@ All notable changes to this project will be documented in this file.
 - CartDecomposition parameter for the distributed structures is now optional
 - template getPos<0>(), substituted by getPos()
 
-## [0.2.1] - 
+## [0.2.1] - 01-04-2016
 
 ### Changed
 - GoogleChart name function changed: AddPointGraph to AddLinesGraph and AddColumsGraph to AddHistGraph